+++ src/app/login/page.tsx
@@ -1,10 +1,6 @@
 import type { Metadata } from "next";
-import { Suspense } from "react";
 
-import { Box, Heading } from "@/components";
-
-import { Inputs } from "./Inputs";
-import { InputsSkeleton } from "./InputsSkeleton";
+import { LoginForm } from "@/features/auth";
 
 export const metadata: Metadata = {
   title: "Login",
@@ -12,18 +8,7 @@ export const metadata: Metadata = {
 
 const Page = () => (
   <div className="flex h-full w-full items-center justify-center self-center">
-    <Box>
-      <form className="flex w-full flex-col">
-        <div className="text-center">
-          <Heading type="title-large">
-            <div className="text-accent-light dark:text-accent-dark">Please login to continue</div>
-          </Heading>
-        </div>
-        <Suspense fallback={<InputsSkeleton />}>
-          <Inputs />
-        </Suspense>
-      </form>
-    </Box>
+    <LoginForm />
   </div>
 );
