+++ src/app/problems/all/[page]/ProblemScore.tsx
@@ -1,7 +1,8 @@
-import { getMemoProblemsScore, getMemoUserInfo } from "@utils/api";
+import { getMemoProblemsScore } from "@utils/api";
 import { cookies } from "next/headers";
 
 import { Score } from "@/components";
+import { getMemoUserInfo } from "@/features/auth";
 
 export const ProblemScoreText = ({ text }: { text: string }) => (
   <p className="font-semibold text-slate-600 dark:text-slate-200">{text}</p>
