+++ src/components/NavigationBar/NavLinks.tsx
@@ -1,12 +1,12 @@
 "use client";
 import { Menu, RadioGroup, Transition } from "@headlessui/react";
-import { logout } from "@utils/actions";
 import { clsx, getCodefunRole, getCodefunRoleTextClass } from "@utils/shared";
 import Image from "next/image";
 import { usePathname, useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
 
 import { ComputerIcon, MoonIcon, SunIcon, UserIcon } from "@/components/icon";
+import { logout } from "@/features/auth";
 import { useAppDispatch, useAppSelector } from "@/hooks";
 import { setScheme, setUser } from "@/store/redux";
