+++ src/features/auth/types/index.ts
@@ -1,0 +1,8 @@
+import type { UserData } from "@/types";
+
+export interface LoginFormState {
+  user: UserData | null;
+  username_messages: string[];
+  password_messages: string[];
+  messages: string[];
+}
