+++ src/shared/utils/actions/index.ts
@@ -1,5 +1,2 @@
-export { authenticate } from "./authenticate";
 export { createProblem } from "./createProblem";
-export { login } from "./login";
-export { logout } from "./logout";
 export { submit } from "./submit";
