+++ src/shared/utils/api/index.ts
@@ -1,9 +1,7 @@
 import "server-only";
 export { getAllProblem } from "./getAllProblem";
 export { getMemoProblemsScore } from "./getMemoProblemsScore";
-export { getMemoUserInfo } from "./getMemoUserInfo";
 export { getProblemCount } from "./getProblemCount";
 export { getProblemInfo } from "./getProblemInfo";
 export { type DetailedScoreInfo, getProblemScore } from "./getProblemScore";
 export { getSubmissionInfo, type SubmissionInfo } from "./getSubmissionInfo";
-export { getUserInfo } from "./getUserInfo";
